+++ src/podofo/main/PdfElement.h
@@ -7,7 +7,6 @@
 #ifndef PDF_ELEMENT_H
 #define PDF_ELEMENT_H
 
-#include "PdfDeclarations.h"
 #include "PdfObject.h"
 
 namespace PoDoFo {